2007-04-17  Sami Wagiaalla  <swagiaal@redhat.com>
        * sessions/SessionManager.java: Added check for trying to add a
        Session which has already been previously added to the manager.
        * sessions/Session.java: Added boolean setter/getter for checking
        whether procs have already been added to this session or not.
        * gladedir/procpop.glade: Left indent the button labeles in Observers
        Dialog.
        * gladedir/frysk_session_manager.glade: Changed the dialog title from
        Frysk Startup Manager to Frysk Session Manager.
        * gladedir/frysk_create_session_druid.glade: Added first introductory
        page to the druid.
        * druid/CreateFryskSessionDruid.java: Added public static variables for
        page numbers.
        Fixed bug with name entry not being set when editing session.
        Fixed bug with incorrect button sensitivity.
        Added first introductory page handling.
        Added a private function and global variables for setting warning
        dialogs.

2007-04-17  Sami Wagiaalla  <swagiaal@redhat.com>

        * eventviewer/TimeLine.java: Fixed bug todo with Terminated threads
        being labeled like regular threads when reselected.
        * UniqueHashMap.java: Added check for removing objects which are not
        in the internal hashTables
        * GuiObject.java: setName now ignores the call if the name is being
        set to the same name.

+++ frysk-gui/frysk/gui/monitor/UniqueHashMap.java
@@ -101,10 +101,14 @@ public class UniqueHashMap{
 	  public void update (Observable observable, Object object)
 	  {
 	    String key = (String) keyHash.get(object);
-	    nameHash.remove(key);
-	    keyHash.remove(object);
+	    if(nameHash.remove(key)==null){
+	      throw new RuntimeException("could not remove key");
+	    }
+	    if(keyHash.remove(object)==null){
+	      throw new RuntimeException("could not remove object");
+	    }
 	    ((GuiObject) object).propertiesChanged.deleteObserver(nameChangedObserver);
-	    
+	
 	    add((GuiObject) object);
 	  }
 	};
